Last_Point_Distribution_Analysis/requirements.txt
streamlit==1.36.0
streamlit-folium==0.21.1
folium>0.16.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.11.3
requests==2.32.3
gitpython==3.1.43
